A simple tool to check Morph scraper output against Wikidata
This is very much a first version of this. It would at least be a good
idea to extract that code that finds the discrepancies and returns them
as structured data, so this could be used either in a command-line tool,
like this, or in generating a report in some other way.

Initially created by Mark Longair (@mhl) and later worked on
with him by pair-programming.