Update output to group into two sections
We want all items that aren't included in the SPARQL query, whether they
are reconciled to a Wikidata identifier or not, to be in a single
section together. Grouping into two sections like this makes it easier
to understand the output, by grouping all records not found in the
SPARQL query together.