Add a rewrite method to TemplateSection
This allows you to pass an object with a 'rewrite' method (taking new
content for the HTML-comment-protected section and parameters it can
base the rewriting on) to TemplateSection.rewrite method, which will
return a new TemplateSection with the rewritten content.

The idea behind returning a new TemplateSection is that we can avoid
any mutable state in objects of that class.