Update prototype to use daff for comparisons
Previously we were comparing two lists of memberships using their
Wikidata ID. This worked fine for the simple case, but didn't cope with
cases where you want to compare on more than one column, for example
when comparing the heads of Government in Wikidata you want to know if
the office holder _or_ the office itself have changed.

At the moment this is all very procedural, but I think it's important to
get it working first, before we start restructuring it.