Remove unused code
These classes are no longer being used, so remove them so they don't
confuse future developers trying to comprehend this code.

These two classes were added when we thought we'd be doing more of the
heavy lifting with morph and reconciling CSVs against morph. But we've
decided that moving that out of band is a simpler approach, then this
library can focus on generating the comparisons and formatting them.