Un-escape SPARQL query before passing to WQS
Pipe characters need to be escaped in order to include them in a
template, but this means that when we pull the SPARQL query out of the
template we need to un-escape them again.

Ideally this would be solved at the mediawiki-page-replaceable_content
level, and that library would pass back un-escaped parameters, but in
the interests of getting things working right now I'm putting in this
simpler fix.

Fixes #40