Write all templates out to Wikidata
Instead of allowing the user to supply options for writing out templates
we instead write _all_ templates out on the initial run, then the user
can customize them as much as they like. If they want to get the default
template back then can delete the contents of the page, save it and then
refresh the prompt.