Split up default and custom templates
Write the default templates out every time the prompt is refreshed. Then
check for a custom template and if it's missing or empty then fallback
to the default template.

This means that adding or removing columns will be reflected in the
default templates. It also keeps the existing behaviour of blanking a
template to go back to the default, but is now doing this check in
Wikidata, rather than in Ruby.