Check that custom templates exist before rendering
We were previously checking if the default templates returned any
contents using #if, but that doesn't handle templates that don't exist,
because they do actually return contents: a red link to create the
missing page.

This change adds an extra #ifexist check to ensure the template exists
before we render it.