Make comparison page viewable directly
Use the full template name, rather than a "relative" (in MediaWiki
terms) path, so that the `/comparison` subpage of the prompt can be
viewed by visiting that page.

This means that when people are watching the page they can just view the
changes on their own. It also means you can go back through the history
of the comparison page and view older versions of it.