Only match Q values if they span the whole cell
We want to replace values like "Q42" with the full template tag, like
"{{Q|42}}", however this was being a bit over-eager and replacing things
that weren't actually Wikidata IDs.

To avoid this problem I've changed the regex so that it only matches a Q
value if it is the entire contents of the cell.