Normalize underscores to spaces to avoid spurious comparison diffs
At the moment, when the prompt is refresh manually, the page_title
parameter will have underscores instead of any spaces. On the other hand
when it is invoked automatically, the page_title parameter has spaces in
it. This means that when you look at diffs in the version history of the
comparison subpage, you'll see spurious differences (spaces to
underscores) when mixing manual and automatic refreshes.

It seems more robust that the DiffOutputGenerator class normalizes the
page_title parameter to one of these forms so it's consistently dealing
with just one; I've chosen to normalize to spaces in this commit.