Move code to transform cell values from Row to Cell
Rather than creating each Cell with an already processed version of its
contents, move the logic for manipulating the contents into the Cell
class instead.

For now this is just a simple migration of the code, but it allows us to
handle each case differently: e.g. by having slightly different logic
for the CellModified case.