Fix rendering of cells with blank value
There was an uncaught regression introduced when the CellValue class was
added that meant cells with an empty string were causing an error
because `nil` was being passed to the `CellValue` constructor, which
causes an error because it inherits from `String`.

This change fixes that by making `CellValue` a plain class. As an added
bonus this also clears up a reek warning about inheriting from core
classes [1].

[1]: https://github.com/troessner/reek/blob/master/docs/Subclassed-From-Core-Class.md