Factor out a Template class
This isn't quite as well-factored as I'd like yet, as the wikitext
method needs to be passed the outer context to work. But it's a step
towards where we want to go.

As there are no tests for this method yet, I introduced it initially as
a 'wikitext2', in parallel with the previous behaviour, and a
  `binding.pry if wikitext != wikitext2`